feat(core): make DualStreamer inputs non nullable
