+++ core/src/main/java/io/github/thibaultbee/streampack/core/streamers/dual/DualStreamer.kt
@@ -36,26 +36,15 @@ import io.github.thibaultbee.streampack.core.elements.sources.video.camera.exten
 import io.github.thibaultbee.streampack.core.elements.sources.video.mediaprojection.MediaProjectionVideoSourceFactory
 import io.github.thibaultbee.streampack.core.elements.utils.RotationValue
 import io.github.thibaultbee.streampack.core.elements.utils.extensions.displayRotation
-import io.github.thibaultbee.streampack.core.elements.utils.extensions.isCompatibleWith
 import io.github.thibaultbee.streampack.core.interfaces.setCameraId
 import io.github.thibaultbee.streampack.core.pipelines.DispatcherProvider
 import io.github.thibaultbee.streampack.core.pipelines.IDispatcherProvider
-import io.github.thibaultbee.streampack.core.pipelines.StreamerPipeline
-import io.github.thibaultbee.streampack.core.pipelines.StreamerPipeline.AudioOutputMode
-import io.github.thibaultbee.streampack.core.pipelines.outputs.encoding.IConfigurableAudioVideoEncodingPipelineOutput
-import io.github.thibaultbee.streampack.core.pipelines.outputs.encoding.IEncodingPipelineOutputInternal
-import io.github.thibaultbee.streampack.core.pipelines.utils.MultiThrowable
+import io.github.thibaultbee.streampack.core.pipelines.inputs.IAudioInput
+import io.github.thibaultbee.streampack.core.pipelines.inputs.IVideoInput
+import io.github.thibaultbee.streampack.core.pipelines.outputs.encoding.IConfigurableVideoEncodingPipelineOutput
 import io.github.thibaultbee.streampack.core.streamers.infos.IConfigurationInfo
 import io.github.thibaultbee.streampack.core.streamers.single.AudioConfig
 import io.github.thibaultbee.streampack.core.streamers.single.VideoConfig
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combineTransform
-import kotlinx.coroutines.flow.merge
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.runBlocking
 
 /**
  * Creates a [DualStreamer] with a default audio source.
@@ -66,6 +55,8 @@ import kotlinx.coroutines.runBlocking
  * @param firstEndpointFactory the [IEndpointInternal.Factory] implementation of the first output. By default, it is a [DynamicEndpointFactory].
  * @param secondEndpointFactory the [IEndpointInternal.Factory] implementation of the second output. By default, it is a [DynamicEndpointFactory].
  * @param defaultRotation the default rotation in [Surface] rotation ([Surface.ROTATION_0], ...). By default, it is the current device orientation.
+ * @param surfaceProcessorFactory the [ISurfaceProcessorInternal.Factory] implementation to use to create the video processor. By default, it is a [DefaultSurfaceProcessorFactory].
+ * @param dispatcherProvider the [IDispatcherProvider] implementation. By default, it is a [DispatcherProvider].
  */
 @RequiresPermission(Manifest.permission.CAMERA)
 suspend fun cameraDualStreamer(
@@ -74,19 +65,22 @@ suspend fun cameraDualStreamer(
     audioSourceFactory: IAudioSourceInternal.Factory? = MicrophoneSourceFactory(),
     firstEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
     secondEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
-    @RotationValue defaultRotation: Int = context.displayRotation
+    @RotationValue defaultRotation: Int = context.displayRotation,
+    surfaceProcessorFactory: ISurfaceProcessorInternal.Factory = DefaultSurfaceProcessorFactory(),
+    dispatcherProvider: IDispatcherProvider = DispatcherProvider()
 ): DualStreamer {
     val streamer = DualStreamer(
-        context,
-        withAudio = true,
-        withVideo = true,
-        firstEndpointFactory,
-        secondEndpointFactory,
-        defaultRotation
+        context = context,
+        firstEndpointFactory = firstEndpointFactory,
+        secondEndpointFactory = secondEndpointFactory,
+        defaultRotation = defaultRotation,
+        surfaceProcessorFactory = surfaceProcessorFactory,
+        dispatcherProvider = dispatcherProvider
     )
+
     streamer.setCameraId(cameraId)
     if (audioSourceFactory != null) {
-        streamer.audioInput!!.setSource(audioSourceFactory)
+        streamer.setAudioSource(audioSourceFactory)
     }
     return streamer
 }
@@ -99,6 +93,8 @@ suspend fun cameraDualStreamer(
  * @param firstEndpointFactory the [IEndpointInternal.Factory] implementation of the first output. By default, it is a [DynamicEndpointFactory].
  * @param secondEndpointFactory the [IEndpointInternal.Factory] implementation of the second output. By default, it is a [DynamicEndpointFactory].
  * @param defaultRotation the default rotation in [Surface] rotation ([Surface.ROTATION_0], ...). By default, it is the current device orientation.
+ * @param surfaceProcessorFactory the [ISurfaceProcessorInternal.Factory] implementation to use to create the video processor. By default, it is a [DefaultSurfaceProcessorFactory].
+ * @param dispatcherProvider the [IDispatcherProvider] implementation. By default, it is a [DispatcherProvider].
  */
 @RequiresApi(Build.VERSION_CODES.Q)
 suspend fun audioVideoMediaProjectionDualStreamer(
@@ -106,18 +102,20 @@ suspend fun audioVideoMediaProjectionDualStreamer(
     mediaProjection: MediaProjection,
     firstEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
     secondEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
-    @RotationValue defaultRotation: Int = context.displayRotation
+    @RotationValue defaultRotation: Int = context.displayRotation,
+    surfaceProcessorFactory: ISurfaceProcessorInternal.Factory = DefaultSurfaceProcessorFactory(),
+    dispatcherProvider: IDispatcherProvider = DispatcherProvider()
 ): DualStreamer {
     val streamer = DualStreamer(
         context = context,
         firstEndpointFactory = firstEndpointFactory,
         secondEndpointFactory = secondEndpointFactory,
-        withAudio = true,
-        withVideo = true,
-        defaultRotation = defaultRotation
+        defaultRotation = defaultRotation,
+        surfaceProcessorFactory = surfaceProcessorFactory,
+        dispatcherProvider = dispatcherProvider
     )
 
-    streamer.videoInput!!.setSource(MediaProjectionVideoSourceFactory(mediaProjection))
+    streamer.setVideoSource(MediaProjectionVideoSourceFactory(mediaProjection))
     streamer.setAudioSource(MediaProjectionAudioSourceFactory(mediaProjection))
     return streamer
 }
@@ -131,6 +129,8 @@ suspend fun audioVideoMediaProjectionDualStreamer(
  * @param firstEndpointFactory the [IEndpointInternal.Factory] implementation of the first output. By default, it is a [DynamicEndpointFactory].
  * @param secondEndpointFactory the [IEndpointInternal.Factory] implementation of the second output. By default, it is a [DynamicEndpointFactory].
  * @param defaultRotation the default rotation in [Surface] rotation ([Surface.ROTATION_0], ...). By default, it is the current device orientation.
+ * @param surfaceProcessorFactory the [ISurfaceProcessorInternal.Factory] implementation to use to create the video processor. By default, it is a [DefaultSurfaceProcessorFactory].
+ * @param dispatcherProvider the [IDispatcherProvider] implementation. By default, it is a [DispatcherProvider].
  */
 suspend fun videoMediaProjectionDualStreamer(
     context: Context,
@@ -138,16 +138,19 @@ suspend fun videoMediaProjectionDualStreamer(
     audioSourceFactory: IAudioSourceInternal.Factory? = MicrophoneSourceFactory(),
     firstEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
     secondEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
-    @RotationValue defaultRotation: Int = context.displayRotation
+    @RotationValue defaultRotation: Int = context.displayRotation,
+    surfaceProcessorFactory: ISurfaceProcessorInternal.Factory = DefaultSurfaceProcessorFactory(),
+    dispatcherProvider: IDispatcherProvider = DispatcherProvider()
 ): DualStreamer {
     val streamer = DualStreamer(
         context = context,
         firstEndpointFactory = firstEndpointFactory,
         secondEndpointFactory = secondEndpointFactory,
-        withAudio = true,
-        withVideo = true,
-        defaultRotation = defaultRotation
+        defaultRotation = defaultRotation,
+        surfaceProcessorFactory = surfaceProcessorFactory,
+        dispatcherProvider = dispatcherProvider
     )
+
     streamer.setVideoSource(MediaProjectionVideoSourceFactory(mediaProjection))
     if (audioSourceFactory != null) {
         streamer.setAudioSource(audioSourceFactory)
@@ -164,6 +167,8 @@ suspend fun videoMediaProjectionDualStreamer(
  * @param firstEndpointFactory the [IEndpointInternal] implementation of the first output. By default, it is a [DynamicEndpoint].
  * @param secondEndpointFactory the [IEndpointInternal] implementation of the second output. By default, it is a [DynamicEndpoint].
  * @param defaultRotation the default rotation in [Surface] rotation ([Surface.ROTATION_0], ...). By default, it is the current device orientation.
+ * @param surfaceProcessorFactory the [ISurfaceProcessorInternal.Factory] implementation to use to create the video processor. By default, it is a [DefaultSurfaceProcessorFactory].
+ * @param dispatcherProvider the [IDispatcherProvider] implementation. By default, it is a [DispatcherProvider].
  */
 suspend fun DualStreamer(
     context: Context,
@@ -171,16 +176,19 @@ suspend fun DualStreamer(
     videoSourceFactory: IVideoSourceInternal.Factory,
     firstEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
     secondEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
-    @RotationValue defaultRotation: Int = context.displayRotation
+    @RotationValue defaultRotation: Int = context.displayRotation,
+    surfaceProcessorFactory: ISurfaceProcessorInternal.Factory = DefaultSurfaceProcessorFactory(),
+    dispatcherProvider: IDispatcherProvider = DispatcherProvider()
 ): DualStreamer {
     val streamer = DualStreamer(
         context = context,
-        withAudio = true,
-        withVideo = true,
         firstEndpointFactory = firstEndpointFactory,
         secondEndpointFactory = secondEndpointFactory,
-        defaultRotation = defaultRotation
+        defaultRotation = defaultRotation,
+        surfaceProcessorFactory = surfaceProcessorFactory,
+        dispatcherProvider = dispatcherProvider
     )
+
     streamer.setAudioSource(audioSourceFactory)
     streamer.setVideoSource(videoSourceFactory)
     return streamer
@@ -192,238 +200,93 @@ suspend fun DualStreamer(
  * For example, you can use it to live stream and record simultaneously.
  *
  * @param context the application context
- * @param withAudio `true` to capture audio. It can't be changed after instantiation.
- * @param withVideo `true` to capture video. It can't be changed after instantiation.
  * @param firstEndpointFactory the [IEndpointInternal] implementation of the first output. By default, it is a [DynamicEndpoint].
  * @param secondEndpointFactory the [IEndpointInternal] implementation of the second output. By default, it is a [DynamicEndpoint].
  * @param defaultRotation the default rotation in [Surface] rotation ([Surface.ROTATION_0], ...). By default, it is the current device orientation.
  * @param surfaceProcessorFactory the [ISurfaceProcessorInternal.Factory] implementation to use to create the video processor. By default, it is a [DefaultSurfaceProcessorFactory].
+ * @param dispatcherProvider the [IDispatcherProvider] implementation. By default, it is a [DispatcherProvider].
  */
-open class DualStreamer(
-    protected val context: Context,
-    val withAudio: Boolean = true,
-    val withVideo: Boolean = true,
+class DualStreamer(
+    context: Context,
     firstEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
     secondEndpointFactory: IEndpointInternal.Factory = DynamicEndpointFactory(),
     @RotationValue defaultRotation: Int = context.displayRotation,
     surfaceProcessorFactory: ISurfaceProcessorInternal.Factory = DefaultSurfaceProcessorFactory(),
-    dispatcherProvider: IDispatcherProvider = DispatcherProvider(),
+    dispatcherProvider: IDispatcherProvider = DispatcherProvider()
 ) : IDualStreamer, IAudioDualStreamer, IVideoDualStreamer {
-    private val coroutineScope = CoroutineScope(dispatcherProvider.default)
-
-    private val pipeline = StreamerPipeline(
-        context,
-        withAudio,
-        withVideo,
-        AudioOutputMode.PUSH,
-        surfaceProcessorFactory,
-        dispatcherProvider
+    private val streamer = DualStreamerImpl(
+        context = context,
+        withAudio = true,
+        withVideo = true,
+        firstEndpointFactory = firstEndpointFactory,
+        secondEndpointFactory = secondEndpointFactory,
+        defaultRotation = defaultRotation,
+        surfaceProcessorFactory = surfaceProcessorFactory,
+        dispatcherProvider = dispatcherProvider
     )
 
-    private val firstPipelineOutput: IEncodingPipelineOutputInternal =
-        runBlocking(dispatcherProvider.default) {
-            pipeline.createEncodingOutput(
-                withAudio, withVideo, firstEndpointFactory, defaultRotation
-            ) as IEncodingPipelineOutputInternal
-        }
+    override val first = streamer.first as IConfigurableVideoEncodingPipelineOutput
+    override val second = streamer.second as IConfigurableVideoEncodingPipelineOutput
 
+    override val throwableFlow = streamer.throwableFlow
+    override val isOpenFlow = streamer.isOpenFlow
+    override val isStreamingFlow = streamer.isStreamingFlow
 
-    /**
-     * First output of the streamer.
-     */
-    override val first = firstPipelineOutput as IConfigurableAudioVideoEncodingPipelineOutput
-
-    private val secondPipelineOutput: IEncodingPipelineOutputInternal =
-        runBlocking(dispatcherProvider.default) {
-            pipeline.createEncodingOutput(
-                withAudio, withVideo, secondEndpointFactory, defaultRotation
-            ) as IEncodingPipelineOutputInternal
-        }
-
-    /**
-     * Second output of the streamer.
-     */
-    override val second = secondPipelineOutput as IConfigurableAudioVideoEncodingPipelineOutput
-
-    override val throwableFlow: StateFlow<Throwable?> = merge(
-        pipeline.throwableFlow,
-        firstPipelineOutput.throwableFlow,
-        secondPipelineOutput.throwableFlow
-    ).stateIn(
-        coroutineScope,
-        SharingStarted.Eagerly,
-        null
-    )
-
-    /**
-     * Whether any of the output is opening.
-     */
-    override val isOpenFlow: StateFlow<Boolean> = combineTransform(
-        firstPipelineOutput.isOpenFlow, secondPipelineOutput.isOpenFlow
-    ) { isOpens ->
-        emit(isOpens.any { it })
-    }.stateIn(
-        coroutineScope,
-        SharingStarted.Eagerly,
-        false
-    )
-
-    /**
-     * Whether any of the output is streaming.
-     */
-    override val isStreamingFlow: StateFlow<Boolean> = pipeline.isStreamingFlow
-
-    /**
-     * Closes the outputs.
-     * Same as calling [first.close] and [second.close].
-     */
-    override suspend fun close() {
-        firstPipelineOutput.close()
-        secondPipelineOutput.close()
-    }
-
-    // SOURCES
-    override val audioInput = pipeline.audioInput
-
-    // PROCESSORS
-    override val videoInput = pipeline.videoInput
+    override val audioInput: IAudioInput = streamer.audioInput
+    override val videoInput: IVideoInput = streamer.videoInput
 
     /**
      * Sets the target rotation.
      *
      * @param rotation the target rotation in [Surface] rotation ([Surface.ROTATION_0], ...)
      */
-    override suspend fun setTargetRotation(@RotationValue rotation: Int) {
-        pipeline.setTargetRotation(rotation)
-    }
+    override suspend fun setTargetRotation(@RotationValue rotation: Int) =
+        streamer.setTargetRotation(rotation)
 
-    /**
-     * Sets audio configuration.
-     *
-     * It is a shortcut for [IConfigurableAudioVideoEncodingPipelineOutput.setAudioCodecConfig].
-     *
-     * @param audioConfig the audio configuration to set
-     */
     @RequiresPermission(Manifest.permission.RECORD_AUDIO)
-    override suspend fun setAudioConfig(audioConfig: DualStreamerAudioConfig) {
-        val throwables = mutableListOf<Throwable>()
+    override suspend fun setAudioConfig(audioConfig: DualStreamerAudioConfig) =
+        streamer.setAudioConfig(audioConfig)
 
-        val firstAudioCodecConfig = firstPipelineOutput.audioCodecConfigFlow.value
-        if ((firstAudioCodecConfig != null) && (!firstAudioCodecConfig.isCompatibleWith(audioConfig.firstAudioConfig))) {
-            firstPipelineOutput.invalidateAudioCodecConfig()
-        }
-        val secondAudioCodecConfig = secondPipelineOutput.audioCodecConfigFlow.value
-        if ((secondAudioCodecConfig != null) && (!secondAudioCodecConfig.isCompatibleWith(
-                audioConfig.secondAudioConfig
-            ))
-        ) {
-            secondPipelineOutput.invalidateAudioCodecConfig()
-        }
+    override suspend fun setVideoConfig(videoConfig: DualStreamerVideoConfig) =
+        streamer.setVideoConfig(videoConfig)
 
-        try {
-            firstPipelineOutput.setAudioCodecConfig(audioConfig.firstAudioConfig)
-        } catch (t: Throwable) {
-            throwables += t
-        }
-        try {
-            audioConfig.secondAudioConfig.let { secondPipelineOutput.setAudioCodecConfig(it) }
-        } catch (t: Throwable) {
-            throwables += t
-        }
-        if (throwables.isNotEmpty()) {
-            if (throwables.size == 1) {
-                throw throwables.first()
-            } else {
-                throw MultiThrowable(throwables)
-            }
-        }
-    }
+    override suspend fun close() = streamer.close()
 
-    /**
-     * Sets video configuration.
-     *
-     * It is a shortcut for [IConfigurableAudioVideoEncodingPipelineOutput.setVideoCodecConfig].
-     * To only set video configuration for a specific output, use [first.setVideoCodecConfig] or
-     * [second.setVideoCodecConfig] outputs.
-     * In that case, you call [first.setVideoCodecConfig] or [second.setVideoCodecConfig] explicitly,
-     * make sure that the frame rate for both configurations is the same.
-     *
-     * @param videoConfig the video configuration to set
-     */
-    override suspend fun setVideoConfig(videoConfig: DualStreamerVideoConfig) {
-        val throwables = mutableListOf<Throwable>()
+    override suspend fun startStream() = streamer.startStream()
 
-        val firstVideoCodecConfig = firstPipelineOutput.videoCodecConfigFlow.value
-        if ((firstVideoCodecConfig != null) && (!firstVideoCodecConfig.isCompatibleWith(videoConfig.firstVideoConfig))) {
-            firstPipelineOutput.invalidateVideoCodecConfig()
-        }
+    override suspend fun stopStream() = streamer.stopStream()
 
-        val secondVideoCodecConfig = secondPipelineOutput.videoCodecConfigFlow.value
-        if ((secondVideoCodecConfig != null) && (!secondVideoCodecConfig.isCompatibleWith(
-                videoConfig.secondVideoConfig
-            ))
-        ) {
-            secondPipelineOutput.invalidateVideoCodecConfig()
-        }
+    override suspend fun release() = streamer.release()
+}
 
-        try {
-            firstPipelineOutput.setVideoCodecConfig(videoConfig.firstVideoConfig)
-        } catch (t: Throwable) {
-            throwables += t
-        }
-        try {
-            secondPipelineOutput.setVideoCodecConfig(videoConfig.secondVideoConfig)
-        } catch (t: Throwable) {
-            throwables += t
-        }
-        if (throwables.isNotEmpty()) {
-            if (throwables.size == 1) {
-                throw throwables.first()
-            } else {
-                throw MultiThrowable(throwables)
-            }
-        }
-    }
-
-    /**
-     * Configures both video and audio settings.
-     *
-     * It must be call when both stream and audio and video capture are not running.
-     *
-     * Use [IConfigurationInfo] to get value limits.
-     *
-     * If video encoder does not support [VideoConfig.level] or [VideoConfig.profile], it fallbacks
-     * to video encoder default level and default profile.
-     *
-     * @param audioConfig Audio configuration to set
-     * @param videoConfig Video configuration to set
-     *
-     * @throws [Throwable] if configuration can not be applied.
-     * @see [DualStreamer.release]
-     */
-    @RequiresPermission(Manifest.permission.RECORD_AUDIO)
-    suspend fun setConfig(
-        audioConfig: DualStreamerAudioConfig, videoConfig: DualStreamerVideoConfig
-    ) {
-        setAudioConfig(audioConfig)
-        setVideoConfig(videoConfig)
-    }
-
-
-    override suspend fun startStream() = pipeline.startStream()
-
-    override suspend fun stopStream() = pipeline.stopStream()
-
-    override suspend fun release() {
-        pipeline.release()
-        coroutineScope.cancel()
-    }
+/**
+ * Configures both video and audio settings.
+ *
+ * It must be call when both stream and audio and video capture are not running.
+ *
+ * Use [IConfigurationInfo] to get value limits.
+ *
+ * If video encoder does not support [VideoConfig.level] or [VideoConfig.profile], it fallbacks
+ * to video encoder default level and default profile.
+ *
+ * @param audioConfig Audio configuration to set
+ * @param videoConfig Video configuration to set
+ *
+ * @throws [Throwable] if configuration can not be applied.
+ * @see [DualStreamer.release]
+ */
+@RequiresPermission(Manifest.permission.RECORD_AUDIO)
+suspend fun DualStreamer.setConfig(
+    audioConfig: DualStreamerAudioConfig, videoConfig: DualStreamerVideoConfig
+) {
+    setAudioConfig(audioConfig)
+    setVideoConfig(videoConfig)
 }
 
 /**
  * Sets audio configuration.
  *
- * It is a shortcut for [IConfigurableAudioVideoEncodingPipelineOutput.setAudioCodecConfig] when both
+ * It is a shortcut for [DualStreamer.setVideoConfig] when both
  * outputs use the same audio configuration.
  *
  * @param audioConfig the audio configuration to set
@@ -436,7 +299,7 @@ suspend fun DualStreamer.setAudioConfig(audioConfig: AudioConfig) {
 /**
  * Sets video configuration.
  *
- * It is a shortcut for [IConfigurableAudioVideoEncodingPipelineOutput.setVideoCodecConfig] when both
+ * It is a shortcut for [DualStreamer.setVideoConfig] when both
  * outputs use the same video configuration.
  *
  * @param videoConfig the video configuration to set
